Move icons to @usevenice/ui package
