+++ apps/web/components/connections/AddDestinationCard.tsx
@@ -1,4 +1,4 @@
-import {AddFilledIcon} from '../icons'
+import {AddFilledIcon} from '@usevenice/ui/icons'
 import {ResourceCard} from '../ResourceCard'
 
 export function AddDestinationCard() {
