+++ packages/ui/icons/DeleteIcon.tsx
