+++ packages/ui/icons/DocsIcon.tsx
