+++ packages/ui/icons/SyncIcon.tsx
